Allow Rails autoloading to work by properly namespacing presenter components.
